Django 1.8.x: update email validation code
Since Django 1.6, django.core.validators.email_re has disappeared.

Now the right way to do email validation is to instead use either the
EmailValidator() class, or the email_validator()
method (default instance of the class).

Replace the email_re import and use code in edumanage/fields.py with an
EmailValidator instance, customized with empty whitelist (no '@localhost'
domain permitted) and custom error message (as it is used in the MultiEmail
field - it should hint which email address failed the validation).