Django 1.8.x: explicitly list all fields on forms
Since Django 1.8, django.forms.ModelForm requires listing either the fields to
include or the fields to include.  Previously, not setting any of the lists
resulted into all fields being included, but this is no longer permitted and
results into an ImproperlyConfigured exception.

This is documented in detail at
https://docs.djangoproject.com/en/1.9/topics/forms/modelforms/#selecting-the-fields-to-use

Fix this by addding the special field name '__all__' to the Meta inner class in all ModelForm subclasses.