Dj1.8: repl longerusername with custom user model
The longerusername package is not really compatible with Django 1.7+ - it only
hooks into South migrations, and Django 1.7+ does not allow the "monkeypatch"
approach that longerusername uses to modify the Auth.User model.

Instead, do the proper thing and define a custom user model as per
https://docs.djangoproject.com/en/1.8/topics/auth/customizing/#specifying-a-custom-user-model

To make the transition easy:
* Create a model that inherits from Auth.AbstractUser, only modifying the username length
* Make the model use the auth_user DB table so it picks up existing data
* Call the model User so that links from other existing tables still work

For existing Django 1.4 databases, the model would just inherit the existing table structure.

For new Django 1.8 databases, the generated initial migrations would create the tables the same way.