Dj1.8: CustomUM: accounts.User: properly swappable
Add property "swappable" to accounts.User Meta class same as is defined in auth.User:

class User(AbstractUser):
    ...

    class Meta(AbstractUser.Meta):
        swappable = 'AUTH_USER_MODEL'

The value of the property is the name of a setting from the Django project's
settings that defines which model class implements a particular feature (the
user model in this case).

With this feature on, only the model class where the name of the class matches
the value of the setting referenced by the property is active.

Without adding this property, and when changing settings.AUTH_USER_MODEL back
to auth.User (either explicitly or by removing the setting and letting it
default to this value), the accounts.User model class would clash with
auth.User - they'd be both adding foreign key relations with the same name to
related classes.