Dj1.8: CustomUM: swappable prop also to migrations
Django 1.8: upgrade: Custom User Model: add swappable property for
accounts.User also to the generated migrations.

(These missed it as they were generated only after the swappable property was
added to the accounts.User model class).

Without the property also being set in the generated migrations, the migrations
framework does not consider the class swapped out and attempts to run the
migrations for this class even when e.g. auth.User is used, clashing with the
migrations for auth.User and breaking the database initialization.